requirements.txt
streamlit
pandas
python-whois
requests
